requirements.txt
openai
langchain
serpapi
python-dotenv
tiktoken
